refactor(numberGenerator): 패키지 명명 규칙에 따른 패키지명 변경
