+++ src/test/java/racingcar/vo/CarsTest.java
@@ -9,8 +9,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import racingcar.numberGenerator.NumberGenerator;
-import racingcar.numberGenerator.RandomNumberGenerator;
+import racingcar.numbergenerator.NumberGenerator;
+import racingcar.numbergenerator.RandomNumberGenerator;
 import racingcar.view.ErrorMessage;
 
 class CarsTest {
